Abtsract prerender requests into another file
